Example of rc.conf script
This is an example of a complete rc.conf script. In the target system the first lines might differ and the parts that need to be added are below the comment. The beginning of your rc.conf might look different based on installation and system options, but to start the delay the 3 lines below the #elements required for the delaysim have to be added.